Add drawing-layer to map image export
